refactor: extract langchain related module and test it in pytest
