+++ regex_inference/inference/engine.py
@@ -11,57 +11,24 @@ TODO:
 import typing
 from typing import List, Optional, Callable, Any
 import re
-import os
-from langchain import PromptTemplate
-from langchain.llms import OpenAI
-from langchain import LLMChain
 from .filter import Filter
-
-
-def make_verbose(func: Callable) -> Callable:
-    def warp(*args: Any, **kwargs: Any) -> Any:
-        args_str = str(args)
-        kwargs_str = str(kwargs)
-        if len(args_str) > 30:
-            args_str = args_str[:10] + '...' + args_str[-10:]
-        if len(kwargs_str) > 30:
-            kwargs_str = kwargs_str[:10] + '...' + kwargs_str[-10:]
-        print(
-            f'[{func.__name__}]',
-            f'START with input -- args: {args_str}; kwargs: {kwargs_str}')
-        result = func(*args, **kwargs)
-        print(f'END [{func.__name__}]')
-        return result
-    return warp
-
+from .chain import Chain
+from ..utils import make_verbose
 
 class Engine:
-    def __init__(self, openai_api_key: Optional[str] = None, temparature: float = 0.8,
+    def __init__(self, openai_api_key: Optional[str] = None, temperature: float = 0.8,
                  mismatch_tolerance: float = 0.1, max_iteration: int = 3, simpify_regex: bool = True, verbose: bool = False):
-        if openai_api_key is None:
-            openai_api_key = os.environ["OPENAI_API_KEY"]
-        self._openai_llm = OpenAI(
-            openai_api_key=openai_api_key,
-            temperature=temparature,
-            model='text-davinci-003',  # https://platform.openai.com/docs/models/gpt-3-5
-            client='regex_inference'
-        )
+        self._chain = Chain(openai_api_key=openai_api_key, temperature=temperature)
         self._mismatch_tolerance = mismatch_tolerance
         self._max_iteration = max_iteration
         self._simpify_regex = simpify_regex
         if verbose:
             self._make_verbose()
-        self._setup_lang_chains()
 
     @typing.no_type_check
     def _make_verbose(self):
         self.run = make_verbose(self.run)
-        self._run_simplify_regex = make_verbose(
-            self._run_simplify_regex)
-        self._run_alter_regex = make_verbose(
-            self._run_alter_regex)
-        self._run_new_inference = make_verbose(
-            self._run_new_inference)
+        self._run_new_inference = make_verbose(self._run_new_inference)
 
     def run(self, patterns: List[str]) -> str:
         regex_list = self.get_regex_sequence(patterns)
@@ -92,7 +59,7 @@ class Engine:
 
     def _run_alter_regex(self, regex: str, patterns: List[str]) -> str:
         for _ in range(self._max_iteration):
-            result = self._regex_alter_chain.run(
+            result = self._chain.alter_regex.run(
                 regex=regex,
                 strings=Engine._convert_patterns_to_prompt(patterns)
             ).strip()
@@ -105,7 +72,7 @@ class Engine:
 
     def _run_simplify_regex(self, regex: str, patterns: List[str]) -> str:
         for _ in range(self._max_iteration):
-            result = self._regex_simplify_chain.run(
+            result = self._chain.simplify_regex.run(
                 regex=regex,
                 strings=Engine._convert_patterns_to_prompt(patterns)
             ).strip()
@@ -118,7 +85,7 @@ class Engine:
 
     def _run_new_inference(self, patterns: List[str]) -> str:
         for _ in range(self._max_iteration):
-            result = self._new_inference_chain.run(
+            result = self._chain.inference_regex.run(
                 Engine._convert_patterns_to_prompt(patterns)
             ).strip()
             try:
@@ -129,105 +96,7 @@ class Engine:
         return result
 
     def explain(self, regex: str) -> None:
-        result = self._regex_explain_chain.run(regex)
+        result = self._chain.explain_regex.run(regex)
         print(result)
 
-    def _setup_lang_chains(self):
-        self._regex_alter_chain = LLMChain(
-            prompt=self.alter_regex_prompt,
-            llm=self._openai_llm
-        )
-        self._new_inference_chain = LLMChain(
-            prompt=self.new_inference_prompt,
-            llm=self._openai_llm
-        )
-        self._regex_simplify_chain = LLMChain(
-            prompt=self.simplify_regex_prompt,
-            llm=self._openai_llm
-        )
-        self._regex_explain_chain = LLMChain(
-            prompt=self.explain_regex_prompt,
-            llm=self._openai_llm
-        )
-
-    @property
-    def new_inference_prompt(self) -> PromptTemplate:
-        template = """Question: Show me the best and shortest regex that can fully match the strings that I provide to you.
-Note that:
-*. The regex should be as short as possible.
-*. Match sure the resulting regex does not have syntax error.
-*. The regex should full match as many strings as possible.
-*. The regex should not match strings that is not provided.
-*. The number of string combinations matching the resulting regex should be as smaller than the number of target strings provided.
-Now, each instance of the strings that should be fully matched is provided line-by-line and wrapped by double quotes as follows:
-{strings}
-
-Note that:
-1. The double quote is not part of the string instance. Ignore the double quote during inferencing the regex.
-2. Provide the resulting regex without wrapping it in quote
-
-The resulting regex is: """
-        prompt = PromptTemplate(
-            template=template,
-            input_variables=['strings']
-        )
-        return prompt
-
-    @property
-    def alter_regex_prompt(self) -> PromptTemplate:
-        template = """Question: Alter the regex "{regex}" such that the following requirements is matched:
-*. The pattern fully match the regex still fully match the regex.
-*. The regex should full match as many strings provided as possible.
-*. The regex should be as short as possible.
-*. The regex should not match strings that is not provided except for those full match the original regex.
-Now, each instance of the strings is provided line-by-line and wrapped by double quotes as follows:
-{strings}
-
-Note that:
-1. The double quote is not part of the string instance. Ignore the double quote during inferencing the regex.
-2. Provide the resulting regex without wrapping it in quote
-
-The resulting altered regex is: """
-        prompt = PromptTemplate(
-            template=template,
-            input_variables=['regex', 'strings']
-        )
-        return prompt
-
-    @property
-    def simplify_regex_prompt(self) -> PromptTemplate:
-        template = """
-Please revise the regex "{regex}"
-such that the following constraint start with *. can be met:
-*. The original regex consists of multiple regex seperated by "|". Try to combine the similar regex.
-*. After combine, the resulting regex should be as short as possible.
-*. The revised regex should still fully match all the strings full matched the original regex
-*. The revised regex should still fully match each of the strings I provided to you.
-Now, each instance of the strings is provided line-by-line and wrapped by double quotes as follows:
-{strings}
-
-
-Note that:
-1. The double quote is not part of the string instance. Ignore the double quote during inferencing the regex.
-2. Provide the resulting regex without wrapping it in quote
-
-The resulting revise regex is:
-"""
-        prompt = PromptTemplate(
-            template=template,
-            input_variables=['regex', 'strings']
-        )
-        return prompt
-
-    @property
-    def explain_regex_prompt(self) -> PromptTemplate:
-        template = """Question: Explain the regex "{regex}" such that
-1. The role of each character in the regex is elaberated.
-2. Provide 5 most interpretive example strings that fullmatch the regex.
-
-The explaination is: """
-        prompt = PromptTemplate(
-            template=template,
-            input_variables=['regex']
-        )
-        return prompt
+
