+++ tests/test_engine.py
@@ -82,19 +82,6 @@ def patterns():
         "0.30",
         "0.4",
         "0.4.6",
-        "0.4.7",
-        "0.4.8",
-        "0.5",
-        "0.6",
-        "0.7",
-        "0.8",
-        "1.0",
-        "1.0.0",
-        "1.0.1",
-        "1.0.15",
-        "1.0.16",
-        "1.0.17",
-        "1.0.18",
         "1.0.2",
         "1.0.20",
         "1.0.20200721",
@@ -103,16 +90,6 @@ def patterns():
         "1.0.20200812.1",
         "1.0.20200812.2",
         "1.0.20200812.3",
-        "1.0.20200812.4",
-        "1.0.20200812.5",
-        "1.0.20200812.6",
-        "1.0.20200817",
-        "1.0.20200820.10",
-        "1.0.20200820.8",
-        "1.0.20200820.9",
-        "1.0.20200820.post2",
-        "1.0.20200820.post4",
-        "1.0.20200820.post5",
         "1.0.20200820.post6",
         "1.0.20200820.post7",
         "1.0.20200821",
@@ -131,107 +108,14 @@ def patterns():
         "1.0.20200825.8",
         "1.0.20200825.9",
         "1.0.20200827",
-        "1.0.20200827.2",
-        "1.0.20200827.3",
-        "1.0.20200827.4",
-        "1.0.20200827.5",
-        "1.0.20200827.6",
-        "1.0.20200827.7",
-        "1.0.20200827.8",
-        "1.0.20200828",
-        "1.0.20200828.2",
-        "1.0.20200828.3",
-        "1.0.20200828.4",
-        "1.0.20200829",
-        "1.0.20200829.1",
-        "1.0.20200829.2",
-        "1.0.20200829.3",
-        "1.0.20200831",
-        "1.0.20200831.2",
-        "1.0.3",
-        "1.0.4",
-        "1.0.5",
-        "1.0.6",
-        "1.0.7",
-        "1.0.8",
-        "1.0.9",
-        "1.1",
-        "1.1.0",
-        "1.1.1",
-        "1.1.10",
-        "1.1.11",
-        "1.1.12",
-        "1.1.13",
-        "1.1.14",
-        "1.1.15",
-        "1.1.16",
-        "1.1.17",
-        "1.1.18",
-        "1.1.19",
-        "1.1.2",
-        "1.1.20",
-        "1.1.21",
-        "1.1.22",
-        "1.1.23",
-        "1.1.24",
-        "1.1.25",
-        "1.1.26",
-        "1.1.27",
-        "1.1.28",
-        "1.1.29",
-        "1.1.3",
-        "1.1.30",
-        "1.1.31",
-        "1.1.32",
-        "1.1.33",
-        "1.1.34",
-        "1.1.35",
-        "1.1.36",
-        "1.1.37",
-        "1.1.38",
-        "1.1.39",
-        "1.1.4",
-        "1.1.40",
-        "1.1.41",
-        "1.1.42",
-        "1.1.43",
         "1.1.44",
         "1.1.45",
         "1.1.46",
         "1.1.47",
         "1.1.48",
-        "1.1.49",
-        "1.1.5",
-        "1.1.50",
-        "1.1.55",
-        "1.1.56",
-        "1.1.57",
-        "1.1.6",
-        "1.1.7",
-        "1.1.8",
-        "1.1.9",
-        "1.11",
-        "1.12",
-        "1.13",
-        "1.14",
-        "1.15",
-        "1.16",
         "1.2.0",
         "1.2.1",
         "1.2.2",
-        "1.2.3",
-        "1.2.4",
-        "1.2.5",
-        "1.2.6",
-        "1.3.0",
-        "1.3.1",
-        "1.3.2",
-        "1.3.3",
-        "1.3.4",
-        "1.3.5",
-        "1.4.0",
-        "2.1.4",
-        "2016.4.0",
         "2016.6.0",
         "2017.1.0",
     ]
